% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sling.R
\name{slingRcpp}
\alias{slingRcpp}
\title{An implementation of Sling algorithm}
\usage{
slingRcpp(X, y, l, max_iter = 1000)
}
\arguments{
\item{X}{A size n*p matrix}

\item{y}{A size n vector}

\item{l}{A number}

\item{max_iter}{A integer number}
}
\value{
A list containing w, itr
}
\description{
An implementation of Sling algorithm
}
