% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sling.R
\name{standardLassoRcpp}
\alias{standardLassoRcpp}
\title{An implementation of Lasso regression via Coordinate Descent}
\usage{
standardLassoRcpp(X, y, l, max_iter = 1000)
}
\arguments{
\item{X}{A size n*p matrix}

\item{y}{A size n vector}

\item{l}{A number}

\item{max_iter}{A integer number}
}
\value{
A list containing w, itr
}
\description{
An implementation of Lasso regression via Coordinate Descent
}
